Completed this project

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModel');
 
 // @desc Get all the contacts
 // @rote GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 // @desc Get contact
 // @rote GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -23,7 +23,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc Create contact
 // @rote POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is: ",req.body);
     const {name, email, phone} = req.body;
@@ -31,18 +31,23 @@ const createContact = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error('All fields are required');
     }
-    const contact = await Contact.create({name, email, phone});
+    const contact = await Contact.create({ name, email, phone, user_id: req.user.id });
     res.status(201).json(contact);
 });
 
 // @desc Update contact
 // @rote PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+
+    if (contact.user_id.toString() !== req.user.id.toString()) {
+        res.status(403);
+        throw new Error('Unauthorized to update this contact');
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, { new: true });
@@ -51,13 +56,19 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete contact
 // @rote DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
     }
+
+    if (contact.user_id.toString() !== req.user.id.toString()) {
+        res.status(403);
+        throw new Error('Unauthorized to delete this contact');
+    }
+
     await Contact.findByIdAndDelete(contact);
     res.status(200).json({ message: 'Contact deleted successfully' });
 });
